Significant rewrite. Better interface + output
We're now using Python's argparse to take commands in,
support the caching of raw MIME messages from Gmail
(so we don't have to repeatedly bother Google for
the copies of the same emails over and over again...),
and better processing of HTML emails.

We've fixed a lot of bugs in CSV output that were
confusing Mathematica. It seems that Mathematica
is unable to ignore newlines that are within quoted
fields, but now we make sure to strip all newlines
and carriage returns in quoted fields anyway, as we
have no use for them.